+++ Tests/Unit/Private/New-PlumberConfig.Tests.ps1
@@ -15,14 +15,14 @@ Describe 'New-PlumberConfig' {
             $config.Tasks.PSScriptAnalyzer.IncludeTests | Should -BeTrue
             $config.Tasks.ContainsKey('Exclude') | Should -BeFalse
             $config.Tasks.Backticks.Exclude | Should -Be @()
-            $config.Tasks.Backticks.EnforceWhen | Should -Be 'Always'
-            $config.Tasks.CodeCoverage.EnforceWhen | Should -Be 'Always'
-            $config.Tasks.ModuleVersion.EnforceWhen | Should -Be 'Always'
-            $config.Tasks.ChangelogUpdated.EnforceWhen | Should -Be 'Always'
-            $config.Tasks.Manifest.EnforceWhen | Should -Be 'Always'
-            $config.Tasks.PublicFunctions.EnforceWhen | Should -Be 'Always'
-            $config.Tasks.FunctionFiles.EnforceWhen | Should -Be 'Always'
-            $config.Tasks.Naming.EnforceWhen | Should -Be 'Always'
+            $config.Tasks.Backticks.RunWhen | Should -Be 'Always'
+            $config.Tasks.CodeCoverage.RunWhen | Should -Be 'Always'
+            $config.Tasks.ModuleVersion.RunWhen | Should -Be 'Always'
+            $config.Tasks.ChangelogUpdated.RunWhen | Should -Be 'Always'
+            $config.Tasks.Manifest.RunWhen | Should -Be 'Always'
+            $config.Tasks.PublicFunctions.RunWhen | Should -Be 'Always'
+            $config.Tasks.FunctionFiles.RunWhen | Should -Be 'Always'
+            $config.Tasks.Naming.RunWhen | Should -Be 'Always'
             $config.Tasks.Local | Should -Be @()
             $config.Tasks.PublicFunctionPrefix.Prefix | Should -BeNullOrEmpty
             $config.Tasks.PublicFunctionPrefix.Exclusions | Should -Be @()
@@ -42,7 +42,7 @@ Describe 'New-PlumberConfig' {
                 Tasks     = @{
                     CodeCoverage = @{
                         Minimum     = 90
-                        EnforceWhen = 'OnRelease'
+                        RunWhen = 'OnRelease'
                     }
                     Local = $null
                     PublicFunctionPrefix = @{
@@ -63,7 +63,7 @@ Describe 'New-PlumberConfig' {
             $config.FileScope | Should -Be 'Changed'
             $config.IncludeModuleFolders | Should -Be @('TaskFunctions')
             $config.Tasks.CodeCoverage.Minimum | Should -Be 90
-            $config.Tasks.CodeCoverage.EnforceWhen | Should -Be 'OnRelease'
+            $config.Tasks.CodeCoverage.RunWhen | Should -Be 'OnRelease'
             $config.Tasks.ContainsKey('Exclude') | Should -BeFalse
             $config.Tasks.Local | Should -Be @()
             $config.Tasks.PublicFunctionPrefix.Prefix | Should -Be 'Thing'